+++ tests/t00079/t00079.cc
@@ -1,0 +1,22 @@
+namespace clanguml {
+namespace t00079 {
+
+struct Base { };
+
+struct D { };
+struct E { };
+struct A : public Base {
+    D d;
+    E *e;
+};
+
+struct B {
+    A *a;
+};
+
+struct C {
+    A a;
+};
+
+}
+}